X en Y opnieuw juist

+++ src/Traject.java
@@ -1,17 +1,17 @@
 public class Traject {
-    private int start_crane_x;
-    private int start_crane_y;
+    private double start_crane_x;
+    private double start_crane_y;
     private int startTime_crane;
-    private int pickup_x;
-    private int pickup_y;
-    private int dropoff_x;
-    private int dropoff_y;
+    private double pickup_x;
+    private double pickup_y;
+    private double dropoff_x;
+    private double dropoff_y;
     private int starttijd;
     private int eindtijd;
     private Kraan kraan;
     private int container_id;
 
-    public Traject(int kraanX, int kraanY, int startKraan, int start_x, int start_y, int eind_x, int eind_y, int starttijd, int eindtijd, Kraan kraan, int id) {
+    public Traject(double kraanX, double kraanY, int startKraan, double start_x, double start_y, double eind_x, double eind_y, int starttijd, int eindtijd, Kraan kraan, int id) {
         this.start_crane_x = kraanX;
         this.start_crane_y = kraanY;
         this.startTime_crane = startKraan;
@@ -25,25 +25,25 @@ public class Traject {
         this.container_id = id;
     }
 
-    public int getPickup_x() { return pickup_x; }
+    public double getPickup_x() { return pickup_x; }
 
     public void setPickup_x(int pickup_x) { this.pickup_x = pickup_x; }
 
-    public int getPickup_y() { return pickup_y; }
+    public double getPickup_y() { return pickup_y; }
 
     public void setPickup_y(int pickup_y) { this.pickup_y = pickup_y; }
 
-    public int getDropoff_x() { return dropoff_x; }
+    public double getDropoff_x() { return dropoff_x; }
 
     public void setDropoff_x(int dropoff_x) { this.dropoff_x = dropoff_x; }
 
-    public int getDropoff_y() { return dropoff_y; }
+    public double getDropoff_y() { return dropoff_y; }
 
     public void setDropoff_y(int dropoff_y) { this.dropoff_y = dropoff_y; }
 
     public Kraan getKraan() { return kraan; }
 
-    public int getStart_crane_x() {
+    public double getStart_crane_x() {
         return start_crane_x;
     }
 
@@ -51,7 +51,7 @@ public class Traject {
         this.start_crane_x = start_crane_x;
     }
 
-    public int getStart_crane_y() {
+    public double getStart_crane_y() {
         return start_crane_y;
     }
